1.1.2: Failure kicking, lang.yml update
This version adds a configurable string to the lang.yml file for when players are kicked due to default authentication setup failures (should this occur for any reason). In addition, when users fail to authenticate, notify messages will now alert anyone with the duoauth.notify permission (e.g. admins) which credentials were incorrect.